Tests/LibWeb/Text/expected/css/insert-rule-keyframes-style-invalidation-counters.txt

FAIL: keyframes insertRule only invalidates animation users (2 invalidations)
